+++ R/aaa.R
@@ -21,7 +21,7 @@ globalVariables(c(
 #' @importFrom DiagrammeR DiagrammeROutput renderDiagrammeR DiagrammeR mermaid
 #' @importFrom stats anova as.formula binomial coef confint cor deviance dnorm glm lm na.omit pnorm predict qnorm sd setNames step update weighted.mean wilcox.test rbinom rlnorm rnorm runif terms quantile
 #' @importFrom utils head tail relist as.relistable combn capture.output write.table
-#' @importFrom stats residuals formula model.matrix pt qt
+#' @importFrom stats residuals formula model.matrix pt qt confint.default family
 #' @importFrom methods is
 #' @importFrom import from
 NULL