+++ tests/testthat/test_stats.R
@@ -1,3 +1,6 @@
+# library(radiant.model)
+#library(testthat)
+
 trim <- function(x) gsub("^\\s+|\\s+$", "", x)
 
 ######### tests ########
@@ -50,8 +53,8 @@ test_that("logistic", {
 test_that("logistic - predict", {
   result <- logistic("titanic", "survived", c("pclass", "sex"))
   res1 <- capture.output(predict(result, pred_cmd = "pclass = levels(pclass); sex = 'female'"))[10] %>% trim()
-  # cat(paste0(res1, "\n"))
-  res2 <- "2nd female      0.779"
+  cat(paste0(res1, "\n"))
+  res2 <- "2nd female      0.779 0.712 0.833"
   expect_equal(res1, res2)
 
   result <- logistic("titanic", "survived", "pclass:sex")
@@ -59,8 +62,8 @@ test_that("logistic - predict", {
   expect_equal(res1, res2)
 
   res1 <- capture.output(predict(result, pred_data = "titanic"))[10] %>% trim()
-  # cat(paste0(res1, "\n"))
-  res2 <- "1st female      0.896"
+  cat(paste0(res1, "\n"))
+  res2 <- "1st female      0.896 0.856 0.926"
   expect_equal(res1, res2)
 })